+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to the LA Network Interface Helper Library will be documente
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
 ## [1.2.10] - 2026-05-27
 ### Fixed
 - On macOS, correctly resolve the user defined name (alias) when multiple services exist for the same interface by using the ServiceOrder priority.